+++ src/constants.rs
@@ -27,3 +27,14 @@ pub const VAR_END_BRACE: char = '}';
 pub const EXPR_START: &str = "{{";
 /// End of an attribute expression to be evaluated e.g. `{{ 2 + 2 }}`
 pub const EXPR_END: &str = "}}";
+
+pub const DEFAULT_FONT_SIZE: f32 = 3.0;
+pub const DEFAULT_FONT_FAMILY: &str = "sans-serif";
+pub const DEFAULT_BACKGROUND: &str = "default";
+pub const DEFAULT_RNG_SEED: u64 = 0;
+pub const DEFAULT_SCALE: f32 = 1.0;
+pub const DEFAULT_BORDER: u16 = 5;
+pub const DEFAULT_LOOP_LIMIT: u32 = 1000;
+pub const DEFAULT_VAR_LIMIT: u32 = 1024;
+pub const DEFAULT_DEPTH_LIMIT: u32 = 100;
+pub const DEFAULT_PATH_REPEAT_LIMIT: u32 = 10000;